+++ app/Campaigns/routes.js
@@ -1,0 +1,6 @@
+const Router = require('express').Router()
+const controller = require('./controller')
+
+Router.get('/', controller.index)
+
+module.exports = Router