+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 ### Recources
 - [x] Users
 - [x] Ongs
-- [ ] Campaigns
+- Campaigns
+  - [x] GET /campaigns
+  - [ ] POST /campaigns
+  - [ ] GET /campaigns/:id
+  - [ ] PUT /campaigns/:id
+  - [ ] DELETE /campaigns/:id
 
 #### Application
 - [ ] resources: add pagination